gamepad (joystick) support
a lot of changes in the godot branch, a massive update. plain SDL
version has gamepad support as well.

godot is really acting up, not sure what's going on but it crashes even
when i'm not touching the editor. fortunately the app itself seems to
work fine. the most annoying bit is that the crashes don't even have any
diagnostic information, completely silent.

it would be sad if i would have to give up godot version because it's so
uncooperative. other gui options are rather sad.

android version is neglected this time